week23/python-typing/pyproject.toml
# https://python-poetry.org/docs/pyproject/
# https://www.python.org/dev/peps/pep-0621/
[tool.poetry]
name = "aucampia.demo.typing"
version = "0.1.0-alpha.0"
description = ""
authors = []
readme = "README.md"
packages = [
    {include = "aucampia", from = "src"}
]

[tool.poetry.scripts]

[tool.poetry.dependencies]
python = "^3.6.2"
click = "^7.1.0"
dataclasses = { version = "^0.8", python = ">=3.6,<3.7" }
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
typer = "^0.3.2"
jsonschema = "^3.2.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
coverage = "^5.5"
flake8 = "^3.9.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^21.4.3"
flake8-isort = "^4.0.0"
invoke = "^1.5.0"
isort = "^5.8.0"
mypy = "^0.812"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-subtests = "^0.5.0"
typing-extensions = "^3.10.0"

[build-system]
requires = ["setuptools","poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
